Added waardelijsten for upcoming release
There are 2 files missing when you compare to 5.0.0: Authorities are not really used at this point, and aren't validated for, so skipped those. Machinery_fuel_types are no longer a thing with the way mobile off road currently works.
Other than that the files are identical to 5.0.0, but can't hurt to have them again with upcoming release.